// src/__generated__/subscripiton.ts
/* tslint:disable */
/* eslint-disable */
// @generated
// This file was automatically generated and should not be edited.

// ====================================================
// GraphQL query operation: subscripiton
// ====================================================

export interface subscripiton_subscriptions {
  __typename: "Podcast";
  id: number;
  title: string;
}

export interface subscripiton {
  subscriptions: subscripiton_subscriptions[];
}